api/requirements.txt
fastapi
uvicorn
spacy
fastapi-cors
motor
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
